refactor: datamodels controller tests split per method.
